+++ services/lleidahacker.py
@@ -2,16 +2,13 @@ import glob
 from fastapi import HTTPException
 
 from models.LleidaHacker import LleidaHacker as ModelLleidaHacker
+from models import TokenData
 
 from schemas.LleidaHacker import LleidaHacker as SchemaLleidaHacker
 
-from security import get_password_hash
+from security import check_image_exists, is_service_token
 
 from sqlalchemy.orm import Session
-
-
-def checkImage(imageId: str):
-    return glob.glob("static/" + imageId + ".*")
 
 
 async def get_all(db: Session):
@@ -34,12 +31,16 @@ async def add_lleidahacker(payload: SchemaLleidaHacker, db: Session):
 
 
 async def update_lleidahacker(userId: int, payload: SchemaLleidaHacker,
-                              db: Session):
-    checkImage(lleidahacker.image_id)
+                              db: Session, data: TokenData):
+    if not data.is_admin:
+        if not data.available or data.type != "lleida_hacker" or data.user_id != userId:
+            raise Exception("Not authorized")
+    # if check_image_exists(lleidahacker.image_id)
     lleidahacker = db.query(ModelLleidaHacker).filter(
         ModelLleidaHacker.id == userId).first()
+    if lleidahacker is None:
+        raise Exception("LleidaHacker not found")
     lleidahacker.name = payload.name
-    lleidahacker.password = payload.password
     lleidahacker.nickname = payload.nickname
     lleidahacker.birthdate = payload.birthdate
     lleidahacker.food_restrictions = payload.food_restrictions
@@ -58,9 +59,29 @@ async def update_lleidahacker(userId: int, payload: SchemaLleidaHacker,
     return lleidahacker
 
 
-async def delete_lleidahacker(userId: int, db: Session):
+async def delete_lleidahacker(userId: int, db: Session, data: TokenData):
+    if not data.is_admin:
+        if not data.available or data.type != "lleida_hacker" or data.user_id != userId:
+            raise Exception("Not authorized")
     lleidahacker = db.query(ModelLleidaHacker).filter(
         ModelLleidaHacker.id == userId).first()
+    if lleidahacker is None:
+        raise Exception("LleidaHacker not found")
     db.delete(lleidahacker)
     db.commit()
+    return lleidahacker
+
+async def set_image(userId: int, image_id: str, db: Session, data: TokenData):
+    if not data.is_admin:
+        if not data.available or data.type != "lleida_hacker" or data.user_id != userId:
+            raise Exception("Not authorized")
+    lleidahacker = db.query(ModelLleidaHacker).filter(
+        ModelLleidaHacker.id == userId).first()
+    if lleidahacker is None:
+        raise Exception("LleidaHacker not found")
+    if not check_image_exists(image_id):
+        raise Exception("Image not found")
+    lleidahacker.image_id = image_id
+    db.commit()
+    db.refresh(lleidahacker)
     return lleidahacker
